backend/requirements.txt
flask
Flask-Cors
rembg
Pillow
onnxruntime
numpy
opencv-python